Moved all word functions to a word.c library
